+++ packages/create-cloudflare/templates/hono/pages/templates/worker-configuration.d.ts
