+++ packages/create-cloudflare/templates/hono/pages/templates/wrangler.jsonc
@@ -2,6 +2,7 @@
   "name": "<TBD>",
   "main": "src/index.ts",
   "compatibility_date": "<TBD>",
+	"pages_build_output_dir": "./dist",
   "observability": {
     "enabled": true
   }
