--- packages/create-cloudflare/templates/hono/snippets/appDeclaration.ts
@@ -1,1 +1,0 @@
-new Hono<{ Bindings: CloudflareBindings }>()
