+++ packages/create-cloudflare/templates/hono/workers/templates/public/index.html
