+++ packages/create-cloudflare/templates/hono/workers/templates/src/index.ts
