+++ packages/create-cloudflare/templates/hono/workers/templates/worker-configuration.d.ts
