+++ packages/create-cloudflare/templates/hono/workers/templates/wrangler.jsonc
